refactor: improve Next.js component structure
